requirements.txt
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.0.76
matplotlib==3.7.2
Pillow==10.0.0
pyyaml==6.0.1
